# docs/swagger.yaml
definitions:
  models.Invoice:
    properties:
      cost:
        example: 1.56
        type: number
      ethAddress:
        example: a 42-char eth address with 0x prefix
        maxLength: 42
        minLength: 42
        type: string
    required:
    - ethAddress
    type: object
...
  routes.GenericSiaFileReq:
    properties:
      publicKey:
        example: a 66-character public key
        maxLength: 66
        minLength: 66
        type: string
      requestBody:
        example: look at description for example
        type: string
      signature:
        description: |-
          signature without 0x prefix is broken into
          R: sig[0:63]
          S: sig[64:127]
        example: a 128 character string created when you signed the request with your
          private key or account handle
        maxLength: 128
        minLength: 128
        type: string
    required:
    - publicKey
    - requestBody
    - signature
    type: object
  routes.InitFileSiaUploadReq:
    properties:
      metadataAsFile:
        type: string
      publicKey:
        example: a 66-character public key
        maxLength: 66
        minLength: 66
        type: string
      requestBody:
        example: look at description for example
        type: string
      signature:
        description: |-
          signature without 0x prefix is broken into
          R: sig[0:63]
          S: sig[64:127]
        example: a 128 character string created when you signed the request with your
          private key or account handle
        maxLength: 128
        minLength: 128
        type: string
    required:
    - metadataAsFile
    - publicKey
    - requestBody
    - signature
    type: object
...
  routes.StatusRes:
    properties:
      status:
        example: status of the request
        type: string
    type: object
...
  routes.accountCreateReq:
    properties:
      publicKey:
        example: a 66-character public key
        maxLength: 66
        minLength: 66
        type: string
      requestBody:
        example: look at description for example
        type: string
      signature:
        description: |-
          signature without 0x prefix is broken into
          R: sig[0:63]
          S: sig[64:127]
        example: a 128 character string created when you signed the request with your
          private key or account handle
        maxLength: 128
        minLength: 128
        type: string
    required:
    - publicKey
    - requestBody
    - signature
    type: object
  routes.accountCreateRes:
    properties:
      expirationDate:
        type: string
      invoice:
        $ref: '#/definitions/models.Invoice'
    required:
    - expirationDate
    type: object
...
  routes.accountGetObj:
    properties:
      apiVersion:
        type: integer
      cost:
        example: 2
        type: number
      createdAt:
        type: string
      ethAddress:
        description: the eth address they will send payment to
        example: a 42-char eth address with 0x prefix
        maxLength: 42
        minLength: 42
        type: string
      expirationDate:
        type: string
      maxFolders:
        example: 2000
        type: integer
      maxMetadataSizeInMB:
        example: 200
        type: integer
      monthsInSubscription:
        example: 12
        type: integer
      storageLimit:
        description: how much storage they are allowed, in GB
        example: 100
        type: integer
      storageLocation:
        description: where their files live, on S3 or elsewhere
        type: integer
      storageUsed:
        description: how much storage they have used, in GB
        example: 30
        type: number
      totalFolders:
        example: 2
        type: integer
      totalMetadataSizeInMB:
        example: 1.245765432
        type: number
      updatedAt:
        type: string
    required:
    - apiVersion
    - ethAddress
    - expirationDate
    - monthsInSubscription
    - storageLimit
    type: object
  routes.accountUnpaidRes:
    properties:
      account:
        $ref: '#/definitions/routes.accountGetObj'
      error:
        example: the error encountered while checking
        type: string
      invoice:
        $ref: '#/definitions/models.Invoice'
      paymentStatus:
        example: paid
        type: string
      stripeData:
        $ref: '#/definitions/routes.stripeDataObj'
    required:
    - account
    type: object
...
  routes.deleteFileRes:
    type: object
...
  routes.getAccountDataReq:
    properties:
      publicKey:
        example: a 66-character public key
        maxLength: 66
        minLength: 66
        type: string
      requestBody:
        example: look at description for example
        type: string
      signature:
        description: |-
          signature without 0x prefix is broken into
          R: sig[0:63]
          S: sig[64:127]
        example: a 128 character string created when you signed the request with your
          private key or account handle
        maxLength: 128
        minLength: 128
        type: string
    required:
    - publicKey
    - requestBody
    - signature
    type: object
...
  routes.stripeDataObj:
    properties:
      amount:
        type: number
      chargeID:
        type: string
      chargePaid:
        type: boolean
      opctTxStatus:
        type: string
      stripePaymentExists:
        type: boolean
      stripeToken:
        type: string
    type: object
...
info:
  contact:
    name: Opacity Staff
    url: https://telegram.me/opacitystorage
  description: Opacity backend for file storage.
  license:
    name: OPACITY LIMITED CODE REVIEW LICENSE LICENSE.md
  termsOfService: https://opacity.io/terms-of-service
  title: Storage Node
  version: "1.0"
paths:
  /api/v1/account-data:
    post:
      consumes:
      - application/json
      description: |-
        check the payment status of an account
        requestBody should be a stringified version of (values are just examples):
        {
        "timestamp": 1557346389
        }
      parameters:
      - description: account payment status check object
        in: body
        name: getAccountDataReq
        required: true
        schema:
          $ref: '#/definitions/routes.getAccountDataReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.accountUnpaidRes'
        "400":
          description: 'bad request, unable to parse request body: (with the error)'
          schema:
            type: string
        "404":
          description: 'no account with that id: (with your accountID)'
          schema:
            type: string
      summary: check the payment status of an account
  /api/v1/accounts:
    post:
      consumes:
      - application/json
      description: |-
        create an account
        requestBody should be a stringified version of (values are just examples):
        {
        "storageLimit": 100,
        "durationInMonths": 12,
        }
      parameters:
      - description: account creation object
        in: body
        name: accountCreateReq
        required: true
        schema:
          $ref: '#/definitions/routes.accountCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.accountCreateRes'
        "400":
          description: 'bad request, unable to parse request body: (with the error)'
          schema:
            type: string
        "503":
          description: 'error encrypting private key: (with the error)'
          schema:
            type: string
      summary: create an account
...
  /api/v2/sia/delete:
    post:
      consumes:
      - application/json
      description: |-
        delete a Sia file
        requestBody should be a stringified version of (values are just examples):
        {
        "fileHandle": "the handle of the file",
        }
      parameters:
      - description: file info object
        in: body
        name: routes.GenericSiaFileReq
        required: true
        schema:
          $ref: '#/definitions/routes.GenericSiaFileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.deleteFileRes'
        "400":
          description: 'bad request, unable to parse request body: (with the error)'
          schema:
            type: string
        "500":
          description: some information about the internal error
          schema:
            type: string
      summary: delete a Sia file
...
  /api/v2/sia/init-upload:
    post:
      consumes:
      - multipart/form-data
      description: |-
        start a Sia file upload
        requestBody should be a stringified version of (values are just examples):
        {
        "fileHandle": "a deterministically created file handle",
        "fileSizeInByte": "200000000000006",
        }
      parameters:
      - description: an object to start a Sia file upload
        in: body
        name: InitFileSiaUploadReq
        required: true
        schema:
          $ref: '#/definitions/routes.InitFileSiaUploadReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.StatusRes'
        "400":
          description: 'bad request, unable to parse request body: (with the error)'
          schema:
            type: string
        "403":
          description: signature did not match
          schema:
            type: string
        "500":
          description: some information about the internal error
          schema:
            type: string
      summary: start a Sia file upload
...
swagger: "2.0"